Renamed core package to bigquery_etl_tools
